wasted 2 hours troubleshooting a ghost bug
It turns out when I renamed my dll a few commits ago, I forgot to manually delete the old one. Both were loading which obviously didn't work for the old one because of the other changes I made, but the error message was misleading because it was pointing to the line number in the old one and I was trying to figure out the problem with that line number in the new one.